// src/models/Juso.ts
export interface jusoProps {
  topSido: string
  topGungu: string
  topDong: string
  bottomSido: string
  bottomGungu: string
  bottomDong: string
}

export interface SidoProps {
  sd: string
  x: number
  y: number
}

export interface GunguProps {
  sgg: string
  x: number
  y: number
}

export interface DongProps {
  umd: string
  x: number
  y: number
}